chore:script rename
