+++ README-template.md
@@ -1,17 +1,17 @@
-# stactools-ephemeral
+# stactools-palsar2-scansar
 
-[![PyPI](https://img.shields.io/pypi/v/stactools-ephemeral)](https://pypi.org/project/stactools-ephemeral/)
+[![PyPI](https://img.shields.io/pypi/v/stactools-palsar2-scansar)](https://pypi.org/project/stactools-palsar2-scansar/)
 
-- Name: ephemeral
-- Package: `stactools.ephemeral`
-- [stactools-ephemeral on PyPI](https://pypi.org/project/stactools-ephemeral/)
+- Name: palsar2-scansar
+- Package: `stactools.palsar2_scansar`
+- [stactools-palsar2-scansar on PyPI](https://pypi.org/project/stactools-palsar2-scansar/)
 - Owner: @githubusername
 - [Dataset homepage](http://example.com)
 - STAC extensions used:
   - [proj](https://github.com/stac-extensions/projection/)
 - Extra fields:
-  - `ephemeral:custom`: A custom attribute
-- [Browse the example in human-readable form](https://radiantearth.github.io/stac-browser/#/external/raw.githubusercontent.com/stactools-packages/ephemeral/main/examples/collection.json)
+  - `palsar2-scansar:custom`: A custom attribute
+- [Browse the example in human-readable form](https://radiantearth.github.io/stac-browser/#/external/raw.githubusercontent.com/stactools-packages/palsar2-scansar/main/examples/collection.json)
 
 A short description of the package and its usage.
 
@@ -23,7 +23,7 @@ A short description of the package and its usage.
 ## Installation
 
 ```shell
-pip install stactools-ephemeral
+pip install stactools-palsar2-scansar
 ```
 
 ## Command-line Usage
@@ -31,10 +31,10 @@ pip install stactools-ephemeral
 Description of the command line functions
 
 ```shell
-stac ephemeral create-item source destination
+stac palsar2-scansar create-item source destination
 ```
 
-Use `stac ephemeral --help` to see all subcommands and options.
+Use `stac palsar2-scansar --help` to see all subcommands and options.
 
 ## Contributing
