+++ environment.yml
@@ -1,4 +1,4 @@
-name: stactools-ephemeral
+name: stactools-palsar2-scansar
 channels:
   - conda-forge
   - defaults
