+++ scripts/test
@@ -25,7 +25,7 @@ if [ "${BASH_SOURCE[0]}" = "${0}" ]; then
             *.py ./**/*.py \
             *.md
 
-        pytest --cov=stactools.ephemeral tests
+        pytest --cov=stactools.palsar2_scansar tests
         coverage xml
     fi
 fi
