+++ setup.cfg
@@ -1,14 +1,14 @@
 [metadata]
-name = stactools-ephemeral
-version = attr: stactools.ephemeral.__version__
+name = stactools-palsar2-scansar
+version = attr: stactools.palsar2_scansar.__version__
 description = PROVIDE DESCRIPTION HERE
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = stac-utils
 author_email = stac@radiant.earth
-url = https://github.com/stactools-packages/ephemeral
+url = https://github.com/stactools-packages/palsar2-scansar
 project_urls =
-    Issues = https://github.com/stactools-packages/ephemeral/issues
+    Issues = https://github.com/stactools-packages/palsar2-scansar/issues
 keywords =
     stactools
     pystac
