+++ src/stactools/palsar2_scansar/__init__.py
@@ -1,7 +1,7 @@
 import stactools.core
 from stactools.cli.registry import Registry
 
-from stactools.ephemeral.stac import create_collection, create_item
+from stactools.palsar2_scansar.stac import create_collection, create_item
 
 __all__ = ["create_collection", "create_item"]
 
@@ -9,9 +9,9 @@ stactools.core.use_fsspec()
 
 
 def register_plugin(registry: Registry) -> None:
-    from stactools.ephemeral import commands
+    from stactools.palsar2_scansar import commands
 
-    registry.register_subcommand(commands.create_ephemeralcmd_command)
+    registry.register_subcommand(commands.create_palsar2scansar_command)
 
 
 __version__ = "0.1.0"
